requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
chromadb>=0.5.0
sentence-transformers>=2.5.1
transformers>=4.40.0
torch>=2.2.0
pypdf>=4.2.0
python-dotenv>=1.0.1
groq>=0.9.0
